pyproject.toml
[project]
name = "chatpdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.2",
    "jinja2>=3.1.6",
    "nltk>=3.9.1",
    "openai>=1.107.3",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
